pyproject.toml
[project]
name = "starter-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-cohere>=0.0.5",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "requests>=2.28.0",
    "sentence-transformers>=2.6.0",
    "tiktoken>=0.5.0",
    "ragas>=0.4.3",
    "datasets>=4.7.0",
    "turbopuffer>=1.18.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint.per-file-ignores]
# Examples use sys.path.append before importing arlc/retrieval
"examples/**/*.py" = ["E402"]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
